Work on the services and controllers and searching with pagination

+++ src/main/java/com/example/cleaningcompanyplanner/assignment/AssignmentRepository.java
@@ -2,5 +2,6 @@ package com.example.cleaningcompanyplanner.assignment;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AssignmentRepository extends JpaRepository<Assignment, Long> {
+public interface AssignmentRepository extends JpaRepository<Assignment, Integer> {
+
 }
